requirements.txt
boto3
selenium ==4.3.0
requests
sqlalchemy
pandas
psycopg2-binary
webdriver-manager
undetected_chromedriver

Flask==1.1.1
gunicorn==20.0.4
chromedriver-binary==79.0.3945.36






